tests: move test into own class and switch to calling via http
The dedicated class is necessary to setup the environment to that the
routes get properly declared (otherwise switching to http in the test
won't work)
