+++ tests/Unit/TypesInSchemas/TypesTest.php
@@ -188,65 +188,6 @@ GRAPHQL;
         self::assertSame($expected, $actual);
     }
 
-    public function testQueriesAndTypesEachInTheirOwnSchema(): void
-    {
-        $this->app['config']->set('graphql.schemas.default', [
-            'query' => [
-                SchemaOne\Query::class,
-            ],
-            'types' => [
-                SchemaOne\Type::class,
-            ],
-        ]);
-        $this->app['config']->set('graphql.schemas.custom', [
-            'query' => [
-                SchemaTwo\Query::class,
-            ],
-            'types' => [
-                SchemaTwo\Type::class,
-            ],
-        ]);
-
-        $query = <<<'GRAPHQL'
-{
-    query {
-        name
-    }
-}
-GRAPHQL;
-
-        $actual = $this->httpGraphql($query);
-
-        $expected = [
-            'data' => [
-                'query' => [
-                    'name' => 'example from schema one',
-                ],
-            ],
-        ];
-        self::assertSame($expected, $actual);
-
-        $query = <<<'GRAPHQL'
-{
-    query {
-        title
-    }
-}
-GRAPHQL;
-        $actual = GraphQL::query($query, null, [
-            'schema' => 'custom',
-        ]);
-
-        $expected = [
-            'data' => [
-                'query' => [
-                    'title' => 'example from schema two',
-                ],
-            ],
-        ];
-        self::assertSame($expected, $actual);
-    }
-
     public function testSameQueryInDifferentSchemasAndTypeGlobal(): void
     {
         $this->app['config']->set('graphql.schemas.default', [
